reference-server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pynacl==1.5.0
pydantic==2.5.0